refactor: #145
응답 dto 분리 및 command 분리
